# oxTrust/api-server/src/main/resources/META-INF/openapi.yaml
openapi: 3.0.1
info:
  title: oxTrust API
  description: This is an API for Gluu Server's oxTrust administrative interface.
    Go to https://gluu.org for more information
  termsOfService: https://gluu.org/gluu-terms-and-conditions/
  contact:
    name: Gluu Support
    url: https://support.gluu.org/
    email: support@gluu.org
  license:
    name: Gluu Support License
    url: https://gluu.org/
  version: 4.5.4
paths:
  /api/v1/configuration/api:
    get:
      summary: Retrieve api configuration
      description: Retrieve api configuration
      operationId: retrieveApiConfiguration
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiConfig'
        500:
          description: Server error
      security:
      - oauth2:
        - https://gluu.org/auth/oxtrust.apiconfig.read
    post:
      summary: Update api configuration
      description: Update api configuration
      operationId: updateApiConfiguration
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiConfig'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiConfig'
        500:
          description: Server error
...
  /api/v1/configuration/idp:
    get:
      summary: Retrieve idp configuration
      description: Retrieve idp configuration
      operationId: retrieveIdpConfiguration
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdpConfig'
        500:
          description: Server error
    post:
      summary: Update idp configuration
      description: Update idp configuration
      operationId: updateIdpConfiguration
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdpConfig'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdpConfig'
        500:
          description: Server error
...
  /api/v1/configuration/oxauth:
    get:
      summary: Retrieve oxauth configuration
      description: Retrieve oxauth configuration
      operationId: retrieveOxauthConfiguration
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OxAuthJsonConfiguration'
        500:
          description: Server error
    put:
      summary: Update json oxauth settings
      description: Updates the oxAuth JSON configuration
      operationId: updateOxauthConfiguration
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OxAuthJsonConfiguration'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OxAuthJsonConfiguration'
        500:
          description: Server error
...
components:
  schemas:
    ApiConfig:
      type: object
      properties:
        apiUmaClientId:
          type: string
        apiUmaClientKeyId:
          type: string
        apiUmaResourceId:
          type: string
        apiUmaScopes:
          type: string
          writeOnly: true
        apiUmaScope:
          type: string
        apiUmaClientKeyStoreFile:
          type: string
        apiUmaClientKeyStorePassword:
          type: string
        oxTrustApiTestMode:
          type: boolean
...
    IdpConfig:
      type: object
      properties:
        idpLdapProtocol:
          type: string
        idpLdapServer:
          type: string
        idpBindDn:
          type: string
        idpBindPassword:
          type: string
        idpUserFields:
          type: string
        idpUrl:
          type: string
        idpSecurityKey:
          type: string
        idpSecurityKeyPassword:
          type: string
        idpSecurityCert:
          type: string
        idp3SigningCert:
          type: string
        idp3EncryptionCert:
          type: string
        shibbolethVersion:
          type: string
        shibboleth3IdpRootDir:
          type: string
        shibboleth3SpConfDir:
          type: string
        shibboleth3FederationRootDir:
          type: string
...
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://{op-hostname}/.../token
          scopes:
            https://gluu.org/auth/oxtrust.apiconfig.read: View Api Config related
              information
            https://gluu.org/auth/oxtrust.apiconfig.write: Manage Api Config related
              information
            https://gluu.org/auth/oxtrust.attribute.read: View Attribute related information
            https://gluu.org/auth/oxtrust.attribute.write: Manage Attribute related
              information
            https://gluu.org/auth/oxtrust.authenticationmethod.read: View Authentication
              Method related information
            https://gluu.org/auth/oxtrust.authenticationmethod.write: Manage Authentication
              Method related information
            https://gluu.org/auth/oxtrust.captchaconfig.read: View Captcha Config
              related information
            https://gluu.org/auth/oxtrust.captchaconfig.write: Manage Captcha Config
              related information
            https://gluu.org/auth/oxtrust.casprotocol.read: View Cas Protocol related
              information
            https://gluu.org/auth/oxtrust.casprotocol.write: Manage Cas Protocol related
              information
            https://gluu.org/auth/oxtrust.certificates.read: View Certificates related
              information
            https://gluu.org/auth/oxtrust.certificates.write: Manage Certificates
              related information
            https://gluu.org/auth/oxtrust.client.read: View Client related information
            https://gluu.org/auth/oxtrust.client.write: Manage Client related information
            https://gluu.org/auth/oxtrust.configuration.read: View Configuration related
              information
            https://gluu.org/auth/oxtrust.configuration.write: Manage Configuration
              related information
            https://gluu.org/auth/oxtrust.customscript.read: View Custom Script related
              information
            https://gluu.org/auth/oxtrust.customscript.write: Manage Custom Script
              related information
            https://gluu.org/auth/oxtrust.group.read: View Group related information
...